+++ lib/widgets/categories/select_color_field.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_colorpicker/flutter_colorpicker.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:lotti/themes/colors.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/color.dart';
 import 'package:lotti/widgets/settings/categories/categories_type_card.dart';
@@ -54,7 +53,7 @@ class _SelectColorFieldState extends State<SelectColorField> {
 
     final color = widget.hexColor != null
         ? colorFromCssHex(widget.hexColor)
-        : secondaryTextColor;
+        : Theme.of(context).colorScheme.outline;
 
     Future<void> onTap() async {
       await showModalBottomSheet<void>(
@@ -100,13 +99,13 @@ class _SelectColorFieldState extends State<SelectColorField> {
         icon: ColorIcon(color),
         hintText: localizations.colorPickerHint,
         hintStyle: style.copyWith(
-          color: secondaryTextColor.withOpacity(0.5),
+          color: Theme.of(context).colorScheme.outline.withOpacity(0.5),
         ),
         suffixIcon: IconButton(
           onPressed: onTap,
           icon: Icon(
             Icons.color_lens_outlined,
-            color: secondaryTextColor,
+            color: Theme.of(context).colorScheme.outline,
             semanticLabel: 'Pick color',
           ),
         ),