+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Colors in dashboard chart selector
 - Colors in entry delete bottom sheet
 - Colors in measurables definition page
+- Tag search text color
 
 ## [0.8.366] - 2023-06-03
 ### Changed: